move access tests to functional
